late night bug fixes

+++ procedural/adduser.php
@@ -3,10 +3,15 @@
 if (isset($_POST['submit'])) {
     $errors = validateInputHasKeys($_POST, ['name', 'email', 'age']);
     if (empty($errors)) {
+        $userExists = checkIsUnique($_POST['email'], $connection);
+        if ($userExists) {
+            errorAlert('User with this email already exists');
+            return;
+        }
         $result = addNewUser($_POST, $connection);
         if ($result) {
             // This is a convenient way to naturally send the user back to the main page.
-            echo '<div class="alert alert-success" role="alert">User added successfully!</div>';
+            successAlert('User added successfully!');
             echo '<script>setTimeout(() => window.location = "index.php?route=getusers", 1000)</script>';
         }
     }
